♻️ refactor: move portal code to features folder
